refactor separate codes
