+++ Kloudy/Kloudy/Model/LocationData.swift
@@ -1,0 +1,14 @@
+//
+//  LocationData.swift
+//  Kloudy
+//
+//  Created by Seulki Lee on 2022/11/20.
+//
+
+import Foundation
+
+struct LocationData {
+    var code: String
+    var city: String
+    var province: String
+}
